+++ cmd/kubeadm/app/phases/certs/renewal/renewal_test.go
@@ -33,7 +33,7 @@ import (
 	k8stesting "k8s.io/client-go/testing"
 	certutil "k8s.io/client-go/util/cert"
 	"k8s.io/kubernetes/cmd/kubeadm/app/phases/certs"
-	"k8s.io/kubernetes/cmd/kubeadm/app/phases/certs/pkiutil"
+	"k8s.io/kubernetes/cmd/kubeadm/app/util/pkiutil"
 	testutil "k8s.io/kubernetes/cmd/kubeadm/test"
 	certtestutil "k8s.io/kubernetes/cmd/kubeadm/test/certs"
 )
